Try to protect against invalid draw counts without crashing
* This can be especially error-prone with indirect draws where the
  parameters are GPU-fed and may point to random memory.

+++ renderdoc/driver/d3d11/d3d11_replay.h
@@ -303,7 +303,7 @@ private:
 
   HighlightCache m_HighlightCache;
 
-  uint32_t m_SOBufferSize = 32 * 1024 * 1024;
+  uint64_t m_SOBufferSize = 32 * 1024 * 1024;
   ID3D11Buffer *m_SOBuffer = NULL;
   ID3D11Buffer *m_SOStagingBuffer = NULL;
   std::vector<ID3D11Query *> m_SOStatsQueries;
