+++ renderdoc/driver/gl/gl_postvs.cpp
@@ -363,7 +363,7 @@ void GLReplay::InitPostVSBuffers(uint32_t eventId)
 
   if(!(drawcall->flags & DrawFlags::UseIBuffer))
   {
-    uint32_t outputSize = drawcall->numIndices * stride;
+    uint64_t outputSize = uint64_t(drawcall->numIndices) * stride;
 
     if(drawcall->flags & DrawFlags::Instanced)
       outputSize *= drawcall->numInstances;
@@ -371,10 +371,11 @@ void GLReplay::InitPostVSBuffers(uint32_t eventId)
     // resize up the buffer if needed for the vertex output data
     if(DebugData.feedbackBufferSize < outputSize)
     {
-      uint32_t oldSize = DebugData.feedbackBufferSize;
+      uint64_t oldSize = DebugData.feedbackBufferSize;
       while(DebugData.feedbackBufferSize < outputSize)
         DebugData.feedbackBufferSize *= 2;
-      RDCWARN("Resizing xfb buffer from %u to %u for output", oldSize, DebugData.feedbackBufferSize);
+      RDCWARN("Resizing xfb buffer from %llu to %llu for output", oldSize,
+              DebugData.feedbackBufferSize);
       gl.glNamedBufferDataEXT(DebugData.feedbackBuffer, DebugData.feedbackBufferSize, NULL,
                               eGL_DYNAMIC_READ);
     }
@@ -484,10 +485,11 @@ void GLReplay::InitPostVSBuffers(uint32_t eventId)
     // resize up the buffer if needed for the vertex output data
     if(DebugData.feedbackBufferSize < outputSize)
     {
-      uint32_t oldSize = DebugData.feedbackBufferSize;
+      uint64_t oldSize = DebugData.feedbackBufferSize;
       while(DebugData.feedbackBufferSize < outputSize)
         DebugData.feedbackBufferSize *= 2;
-      RDCWARN("Resizing xfb buffer from %u to %u for output", oldSize, DebugData.feedbackBufferSize);
+      RDCWARN("Resizing xfb buffer from %llu to %llu for output", oldSize,
+              DebugData.feedbackBufferSize);
       gl.glNamedBufferDataEXT(DebugData.feedbackBuffer, DebugData.feedbackBufferSize, NULL,
                               eGL_DYNAMIC_READ);
     }
@@ -1034,10 +1036,10 @@ void GLReplay::InitPostVSBuffers(uint32_t eventId)
       // resize up the buffer if needed for the vertex output data
       if(DebugData.feedbackBufferSize < maxOutputSize)
       {
-        uint32_t oldSize = DebugData.feedbackBufferSize;
+        uint64_t oldSize = DebugData.feedbackBufferSize;
         while(DebugData.feedbackBufferSize < maxOutputSize)
           DebugData.feedbackBufferSize *= 2;
-        RDCWARN("Conservatively resizing xfb buffer from %u to %u for output", oldSize,
+        RDCWARN("Conservatively resizing xfb buffer from %llu to %llu for output", oldSize,
                 DebugData.feedbackBufferSize);
         gl.glNamedBufferDataEXT(DebugData.feedbackBuffer, DebugData.feedbackBufferSize, NULL,
                                 eGL_DYNAMIC_READ);
